+++ src/service/LinkService.ts
@@ -45,6 +45,7 @@ export type LinkParams = {
     unlink: boolean;
     useKnownRestGateways: boolean;
     ready?: boolean;
+    customPreset?: string;
     removeOldLinked?: boolean; //TEST ONLY!
 };
 
@@ -86,6 +87,7 @@ export class LinkService implements TransactionFactory {
     public async run(passedPresetData?: ConfigPreset | undefined, passedAddresses?: Addresses | undefined): Promise<void> {
         const presetData = passedPresetData ?? this.configLoader.loadExistingPresetData(this.params.target, this.params.password);
         const addresses = passedAddresses ?? this.configLoader.loadExistingAddresses(this.params.target, this.params.password);
+        const customPreset = this.configLoader.loadCustomPreset(this.params.customPreset, this.params.password);
         logger.info(`${this.params.unlink ? 'Unlinking' : 'Linking'} nodes`);
 
         await new AnnounceService().announce(
@@ -94,7 +96,7 @@ export class LinkService implements TransactionFactory {
             this.params.useKnownRestGateways,
             this.params.ready,
             this.params.target,
-            presetData,
+            this.configLoader.mergePresets(presetData, customPreset),
             addresses,
             this,
         );